Add files via upload
Abstract - Being a global issue, it might look smaller in size, but the littering of cigarette butts (CBs) causes a damaging impact on the environment. It is said that trillions of cigarettes that are produced across globe annually lead to thousands of kilograms of hazardous waste getting into the ground. Most of the CB are made up of slow degradable substance called cellulose acetate. It takes 12 to 15 years for these substances to decompose. During the process of decomposition, it releases toxic chemicals causing serious damage to the environment. To make the best use of these CBs we have come up with an idea of converting these CBs into heat insulation materials that can be used in a safe way after thorough processing and extraction of chemical substances present in them. Our idea is aligned with the SDGs announced by UNO. It falls under SDG 12 - Responsible Consumption and Production. Target 12.5: "By 2030, substantially reduce waste generation through prevention, reduction, recycling, and reuse.". Objective of our idea is to provide a clean and safe environment along with creation of employment opportunity for the Homeless and refugees improving the per capita of the country.